FiraCode Nerd font and fixed sdkman and nvm
sdkman and nvm will only add to the path once. zshrc was being executed once in iTerm was opened and executed a second time with tmux. Since sdkman and nvm were in the path from when iTerm was open it would not add them to the path when tmux initialized. This put sdkman and nvm at the end of the path rather than it being first in the path. Added a check to make sure tmux was initialized so I could add them to the beginning to the path.